refactor: ReservationPolicy -> ReservationCreatePolicy
