+++ README.md
@@ -30,13 +30,14 @@
 - [ ] 추가된 테마와 스케줄에 따라 기존에 구현된 예약을 리팩터링
 - [ ] 예약과 관련된 예외 상황을 처리
     - [x] 예약 등록을 하려는 날짜와 시간에 이미 등록되어있으면 예약을 등록을 실패한다.
-    - [ ] 예약이 없는 경우에도 빈 응답 값을 응답한다.
+    - [x] 예약이 없는 경우에도 빈 응답 값을 응답한다.
     - [ ] 예약이 존재하지 않는 경우 예약 삭제를 실패한다.
+    - [ ] 예약과 관계있는 스케줄 혹은 테마는 수정 및 삭제가 불가능하다
+
 - [ ] 스케쥴과 관련된 예외
     - [x] 테마는 스케쥴을 다르게 설정할 수 있어야한다.
     - [x] 스케쥴 저장 시 테마가 존재하지 않으면 예외가 발생한다.
     - [x] 스케쥴 조회시 테마가 존재하지 않으면 예외가 발생한다.
-    - [ ] 예약과 관계있는 스케줄 혹은 테마는 수정 및 삭제가 불가능하다
 
 ### 프로그래밍 요구사항
