+++ src/main/java/nextstep/Main.java
@@ -3,8 +3,8 @@ package nextstep;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Scanner;
-import nextstep.application.reservation.ReservationCreateValidation;
-import nextstep.application.reservation.ReservationPolicy;
+import nextstep.application.reservation.ReservationCreateExistValidation;
+import nextstep.application.reservation.ReservationCreatePolicy;
 import nextstep.application.reservation.RoomEscapeService;
 import nextstep.application.reservation.dto.Reservation;
 import nextstep.domain.reservation.repository.MemoryReservationRepository;
@@ -19,7 +19,7 @@ public class Main {
   public static void main(String[] args) {
     Scanner scanner = new Scanner(System.in);
     RoomEscapeService service = new RoomEscapeService(new MemoryReservationRepository(),
-        new ReservationPolicy(List.of(new ReservationCreateValidation(new MemoryReservationRepository()))));
+        new ReservationCreatePolicy(List.of(new ReservationCreateExistValidation(new MemoryReservationRepository()))));
 
     while (true) {
       System.out.println("메뉴를 선택하세요.");
