+++ src/main/java/nextstep/application/reservation/ReservationCreateValidation.java
@@ -1,28 +1,10 @@
 package nextstep.application.reservation;
 
-import java.time.LocalTime;
-import lombok.RequiredArgsConstructor;
 import nextstep.application.reservation.dto.Reservation;
-import nextstep.domain.reservation.repository.ReservationRepository;
-import org.springframework.stereotype.Component;
 
-@Component
-@RequiredArgsConstructor
-public class ReservationCreateValidation implements ReservationValidation {
+public interface ReservationCreateValidation {
 
-  private final ReservationRepository repository;
+  void checkValid(Reservation reservation);
 
-  @Override
-  public void checkValid(Reservation reservation) {
-    var entity = repository.findReservationsByDateAndTime(reservation.date(),
-        LocalTime.parse(reservation.time()));
-    if (entity.isPresent()) {
-      throw new IllegalArgumentException("이미 해당 날짜, 시간에 예약이 존재합니다.");
-    }
-  }
-
-  @Override
-  public int priority() {
-    return 1;
-  }
+  int priority();
 }
