--- src/main/java/nextstep/application/reservation/ReservationValidation.java
@@ -1,10 +1,0 @@
-package nextstep.application.reservation;
-
-import nextstep.application.reservation.dto.Reservation;
-
-public interface ReservationValidation {
-
-  void checkValid(Reservation reservation);
-
-  int priority();
-}
